+++ controllers/image.js
@@ -23,6 +23,7 @@ const Item = require("../model/item");
 const Friend = require("../model/friend");
 const Notification = require("../model/notification");
 const Workgroup = require("../model/workgroup");
+const database = require("../util/database");
 
 const Attributes = require("../model/attributes");
 
@@ -428,32 +429,59 @@ module.exports.getAcceptImage = (req, res, next) => {
   if (!notificationId)
     return StatusResponse(res, 421, "No notification Id provided");
 
-  Notification.findByPk(notificationId).then(async (foundNotification) => {
-    if (!foundNotification)
-      return StatusResponse(res, 404, "Notification not found");
+  database
+    .transaction(async (transaction) => {
+      const foundNotification = await Notification.findByPk(notificationId, {
+        transaction: transaction,
+        lock: transaction.LOCK.UPDATE,
+      });
 
-    if (foundNotification.userId !== req.authUserId)
-      return StatusResponse(
-        res,
-        421,
-        "Notification is not for the current user"
-      );
+      if (!foundNotification)
+        return StatusResponse(res, 404, "Notification not found");
 
-    const imageShareNotification = await getStatusIdByName("Share Image");
-    if (foundNotification.type !== imageShareNotification)
-      return StatusResponse(res, 421, "Notification is not an image share");
+      if (foundNotification.userId !== req.authUserId)
+        return StatusResponse(
+          res,
+          421,
+          "Notification is not for the current user"
+        );
 
-    const imageToShare = foundNotification.objectId;
+      const imageShareNotification = await getStatusIdByName("Share Image");
+      if (foundNotification.type !== imageShareNotification)
+        return StatusResponse(res, 421, "Notification is not an image share");
 
-    Helper.copyImage(imageToShare, req.authUserId)
-      .then((copiedImage) => {
-        if (!copiedImage)
-          return StatusResponse(res, 500, "Unable to copy image");
+      const friendStatus = await getStatusIdByName("Friend");
+      const activeFriendship = await Friend.findOne({
+        where: {
+          userId: foundNotification.senderId,
+          friendId: req.authUserId,
+          statusId: friendStatus,
+        },
+        transaction: transaction,
+      });
 
-        foundNotification.destroy().then(() => {
-          return StatusResponse(res, 200, "OK");
-        });
-      })
-      .catch((err) => next(err));
-  });
+      if (!activeFriendship)
+        return StatusResponse(res, 421, "Share is no longer valid");
+
+      const imageToShare = foundNotification.objectId;
+      const copiedImage = await Helper.copyImage(imageToShare, req.authUserId, {
+        transaction: transaction,
+      });
+
+      if (!copiedImage)
+        return StatusResponse(res, 500, "Unable to copy image");
+
+      await foundNotification.destroy({ transaction: transaction });
+      return { ok: true, userId: req.authUserId, notificationId: notificationId };
+    })
+    .then(async (result) => {
+      if (!result || result.ok !== true) return result;
+
+      await Notify.emitNotificationEvent(result.userId, "deleted", {
+        notificationId: result.notificationId,
+        notification: null,
+      });
+      return StatusResponse(res, 200, "OK");
+    })
+    .catch((err) => next(err));
 };