+++ controllers/list.js
@@ -22,6 +22,7 @@ const Election = require("../model/election");
 const Friend = require("../model/friend");
 const Notification = require("../model/notification");
 const Workgroup = require("../model/workgroup");
+const database = require("../util/database");
 
 const Attributes = require("../model/attributes");
 
@@ -595,31 +596,59 @@ module.exports.getAcceptList = (req, res, next) => {
   if (!notificationId)
     return StatusResponse(res, 421, "No notification Id provided");
 
-  Notification.findByPk(notificationId).then(async (foundNotification) => {
-    if (!foundNotification)
-      return StatusResponse(res, 404, "Notification not found");
+  database
+    .transaction(async (transaction) => {
+      const foundNotification = await Notification.findByPk(notificationId, {
+        transaction: transaction,
+        lock: transaction.LOCK.UPDATE,
+      });
 
-    if (foundNotification.userId !== req.authUserId)
-      return StatusResponse(
-        res,
-        421,
-        "Notification is not for the current user"
-      );
+      if (!foundNotification)
+        return StatusResponse(res, 404, "Notification not found");
 
-    const listShareNotification = await getStatusIdByName("Share List");
-    if (foundNotification.type !== listShareNotification)
-      return StatusResponse(res, 421, "Notification is not a list share");
+      if (foundNotification.userId !== req.authUserId)
+        return StatusResponse(
+          res,
+          421,
+          "Notification is not for the current user"
+        );
 
-    const listToShare = foundNotification.objectId;
+      const listShareNotification = await getStatusIdByName("Share List");
+      if (foundNotification.type !== listShareNotification)
+        return StatusResponse(res, 421, "Notification is not a list share");
 
-    Helper.copyList(listToShare, req.authUserId)
-      .then((copiedList) => {
-        if (!copiedList) return StatusResponse(res, 500, "Unable to copy list");
+      const friendStatus = await getStatusIdByName("Friend");
+      const activeFriendship = await Friend.findOne({
+        where: {
+          userId: foundNotification.senderId,
+          friendId: req.authUserId,
+          statusId: friendStatus,
+        },
+        transaction: transaction,
+      });
 
-        foundNotification.destroy().then(() => {
-          return StatusResponse(res, 200, "OK");
-        });
-      })
-      .catch((err) => next(err));
-  });
+      if (!activeFriendship)
+        return StatusResponse(res, 421, "Share is no longer valid");
+
+      const listToShare = foundNotification.objectId;
+      const copiedList = await Helper.copyList(listToShare, req.authUserId, {
+        transaction: transaction,
+      });
+
+      if (!copiedList)
+        return StatusResponse(res, 500, "Unable to copy list");
+
+      await foundNotification.destroy({ transaction: transaction });
+      return { ok: true, userId: req.authUserId, notificationId: notificationId };
+    })
+    .then(async (result) => {
+      if (!result || result.ok !== true) return result;
+
+      await Notify.emitNotificationEvent(result.userId, "deleted", {
+        notificationId: result.notificationId,
+        notification: null,
+      });
+      return StatusResponse(res, 200, "OK");
+    })
+    .catch((err) => next(err));
 };